services/synapse/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
neo4j==5.18.0
redis>=5.0.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
structlog==24.1.0
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
pytest-cov==4.1.0
ruff==0.3.2
mypy==1.9.0
pm4py==2.7.11.8
pandas==2.2.1
rdflib==7.0.0
